--- .eslintrc
@@ -1,29 +1,0 @@
-{
-    "extends": [
-      "eslint:recommended"
-    ],
-  
-    "env": {
-      "browser": false,
-      "es6": true,
-      "node": true
-    },
-  
-    "parserOptions":{
-      "ecmaVersion": "latest",
-      "sourceType": "module",
-      "ecmaFeatures": {
-        "jsx": true,
-        "modules": true,
-        "experimentalObjectRestSpread": true
-      }
-    },
-  
-    "globals": {
-      "document": false,
-      "navigator": false,
-      "window": false
-    },
-
-    "plugins": ["react"]
-  }
