+++ .github/workflows/unit-tests.yml
@@ -12,7 +12,7 @@ jobs:
       - name: Install Node.js
         uses: actions/setup-node@v3
         with:
-          node-version: 16
+          node-version: 14
 
       - name: Install dependencies
         run: npm i
