--- index.js
@@ -1,5 +1,0 @@
-import { IfElse, OnTrue, OnFalse } from "./src/IfElse/index.js";
-import { Switch, Case, Satisfies } from "./src/Switch/index.js";
-import { If, When, Unless } from "./src/Conditionals/index.js";
-
-export { IfElse, Switch, Case, Satisfies, OnTrue, OnFalse, If, When, Unless };
