--- jsconfig.json
@@ -1,11 +1,0 @@
-{
-    "compilerOptions": {
-        "baseUrl": "./",
-        "paths": {
-            "/**": [
-                "./src/**/index.js"
-            ]
-        }
-    },
-    "exclude": ["node_modules"]
-}
