+++ lib/index.d.ts
@@ -1,0 +1,30 @@
+export function OnFalse({ condition, children }: {
+    condition: any;
+}): any;
+export function OnTrue({ condition, children }: {
+    condition: any;
+}): any;
+export function If({ condition, children }: {
+    condition: any;
+}): any;
+export function When({ condition, children }: {
+    condition: any;
+}): any;
+export function Unless({ condition, children }: {
+    condition: any;
+}): any;
+export function IfElse({ condition, children }: {
+    condition: any;
+}): any;
+export function Switch({ condition, children }: {
+    condition: any;
+}): any;
+export function Case({ value, condition, children }: {
+    value: any;
+    condition: any;
+}): any;
+export function Satisfies({ predicate, condition, children }: {
+    predicate: (arg0: any) => boolean;
+    condition: any;
+}): any;
+//# sourceMappingURL=index.d.ts.map
