--- package2.json
@@ -1,37 +1,0 @@
-{
-  "name": "conditional-jsx",
-  "version": "0.9.6",
-  "description": "React components for conditional statements.",
-  "main": "index.js",
-  "type": "module",
-  "repository": "https://github.com/JuanGalilea/conditional-jsx",
-  "bugs": {
-    "url": "https://github.com/JuanGalilea/conditional-jsx/issues"
-  },
-  "author": "JuanGalilea (https://github.com/JuanGalilea)",
-  "files": ["./index.js"],
-  "license": "MIT",
-  "keywords": [
-    "react"
-  ],
-  "scripts": {
-    "test": "vitest run",
-    "coverage": "vitest run --coverage",
-    "start": "node index.js"
-  },
-  "dependencies": {
-    "react": "^18.2.0"
-  },
-  "sideEffects": false,
-  "engines": {
-    "node": "16.x"
-  },
-  "devDependencies": {
-    "@vitest/coverage-c8": "^0.24.3",
-    "babel-eslint": "^10.1.0",
-    "cheerio": "1.0.0-rc.12",
-    "eslint": "^8.25.0",
-    "react-dom": "^18.2.0",
-    "vitest": "^0.24.3"
-  }
-}
