--- src/Conditionals/OnFalse/index.js
@@ -1,10 +1,0 @@
-/**
- * Renders when its IfElse parent condition is falsy
- * - Synonyms: Unless
- * - See also: If, When, OnTrue
- * @function OnFalse
- * @param {object} props
- * @param {*} props.condition Value to check
- */
-export const OnFalse = ({ condition, children }) =>
-    condition ? null : children;
