+++ utils/testUtils.js
@@ -2,4 +2,3 @@ import { renderToStaticMarkup } from "react-dom/server";
 import { load } from "cheerio";
 
 export const render = (component) => load(renderToStaticMarkup(component));
-
